Reorganized files and solution structure.
