+++ Beta.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beta", "src\Beta\Beta.csproj", "{62636448-725A-4AA2-9329-C9FBDFE38A1C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Beta", "src\Beta\Beta.csproj", "{62636448-725A-4AA2-9329-C9FBDFE38A1C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A7B779DC-0CBE-4056-974C-1AE30F3BFB63}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{21FD2E4F-2787-4158-9AE7-6B9F5CD646AC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beta.Tests", "test\Beta.Tests\Beta.Tests.csproj", "{09C7A6A2-0493-4794-A363-60D3807CCEB3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Beta.Tests", "test\Beta.Tests\Beta.Tests.csproj", "{09C7A6A2-0493-4794-A363-60D3807CCEB3}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{82FC7197-DD53-4163-9321-A5328015E746}"
 	ProjectSection(SolutionItems) = preProject
@@ -17,15 +17,17 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beta.Shouldly", "src\Beta.Shouldly\Beta.Shouldly.csproj", "{AEB8BE40-3343-49D1-9169-77481FC03F67}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Beta.Shouldly", "src\Beta.Shouldly\Beta.Shouldly.csproj", "{AEB8BE40-3343-49D1-9169-77481FC03F67}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beta.FakeItEasy", "src\Beta.FakeItEasy\Beta.FakeItEasy.csproj", "{496AD536-AD06-4DF0-A74B-63EE2165DED2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Beta.FakeItEasy", "src\Beta.FakeItEasy\Beta.FakeItEasy.csproj", "{496AD536-AD06-4DF0-A74B-63EE2165DED2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Beta.TestAdapter", "src\Beta.TestAdapter\Beta.TestAdapter.csproj", "{F9DB7757-6554-4F7C-A96C-4072A4FA4714}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Beta.TestAdapter", "src\Beta.TestAdapter\Beta.TestAdapter.csproj", "{F9DB7757-6554-4F7C-A96C-4072A4FA4714}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "mocks", "mocks", "{E33B7854-05E6-449D-872C-9767B8B67F7B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mock-assembly", "src\mock-assembly\mock-assembly.csproj", "{F1B6F762-6C30-4307-AA27-E82842729E80}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "mock-assembly", "src\mock-assembly\mock-assembly.csproj", "{F1B6F762-6C30-4307-AA27-E82842729E80}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "extensions", "extensions", "{C38E9E5B-E8B3-4015-A2CE-5B5FE9ABE2B2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -64,11 +66,12 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{62636448-725A-4AA2-9329-C9FBDFE38A1C} = {A7B779DC-0CBE-4056-974C-1AE30F3BFB63}
 		{09C7A6A2-0493-4794-A363-60D3807CCEB3} = {21FD2E4F-2787-4158-9AE7-6B9F5CD646AC}
-		{AEB8BE40-3343-49D1-9169-77481FC03F67} = {A7B779DC-0CBE-4056-974C-1AE30F3BFB63}
-		{496AD536-AD06-4DF0-A74B-63EE2165DED2} = {A7B779DC-0CBE-4056-974C-1AE30F3BFB63}
+		{AEB8BE40-3343-49D1-9169-77481FC03F67} = {C38E9E5B-E8B3-4015-A2CE-5B5FE9ABE2B2}
+		{496AD536-AD06-4DF0-A74B-63EE2165DED2} = {C38E9E5B-E8B3-4015-A2CE-5B5FE9ABE2B2}
 		{F9DB7757-6554-4F7C-A96C-4072A4FA4714} = {A7B779DC-0CBE-4056-974C-1AE30F3BFB63}
 		{E33B7854-05E6-449D-872C-9767B8B67F7B} = {A7B779DC-0CBE-4056-974C-1AE30F3BFB63}
 		{F1B6F762-6C30-4307-AA27-E82842729E80} = {E33B7854-05E6-449D-872C-9767B8B67F7B}
+		{C38E9E5B-E8B3-4015-A2CE-5B5FE9ABE2B2} = {A7B779DC-0CBE-4056-974C-1AE30F3BFB63}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {75514120-3808-42A3-A9F6-75B4A903FCCF}
