+++ src/Beta/Internal/Discovery/DefaultTestAssemblyExplorer.cs
@@ -1,6 +1,8 @@
 using System.Reflection;
+using Beta.Sdk.Abstractions;
+using Beta.Sdk.Interfaces;
 
-namespace Beta.Discovery;
+namespace Beta.Internal.Discovery;
 
 /// <summary>
 ///     Defines the default assembly explorer.
