+++ src/Beta/Internal/Discovery/DefaultTestCaseDiscoverer.cs
@@ -1,6 +1,9 @@
 using System.Reflection;
+using Beta;
+using Beta.Sdk.Abstractions;
+using Beta.Sdk.Interfaces;
 
-namespace Beta.Discovery;
+namespace Beta.Internal.Discovery;
 
 /// <summary>
 ///     Defines the default test case discoverer.
