+++ src/Beta/Sdk/Interfaces/ITestDiscoverer.cs
@@ -1,4 +1,6 @@
-﻿namespace Beta.Discovery;
+﻿using Beta.Sdk.Abstractions;
+
+namespace Beta.Sdk.Interfaces;
 
 /// <summary>
 ///     Defines a discoverer that can find tests in a given type.
