+++ test/Beta.Tests/Demos/DemoTests.cs
@@ -1,5 +1,6 @@
-﻿using Beta.Discovery;
-using Beta.Execution;
+﻿using Beta.Internal.Discovery;
+using Beta.Internal.Execution;
+using Beta.Sdk.Interfaces;
 using Xunit.Abstractions;
 
 namespace Beta.Tests.Demos;
